# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.AddScheduleRequest:
    properties:
      animalId:
        type: string
      feedingTime:
        type: string
      foodType:
        type: string
    type: object
  model.Animal:
    properties:
      ID:
        type: string
      birthDate:
        type: string
      enclosureID:
        type: string
      favoriteFood:
        $ref: '#/definitions/model.Food'
      gender:
        $ref: '#/definitions/model.Gender'
      healthStatus:
        type: string
      name:
        type: string
      species:
        $ref: '#/definitions/model.Species'
    type: object
  model.Enclosure:
    properties:
      ID:
        type: string
      animalsID:
        items:
          type: string
        type: array
      currentCount:
        type: integer
      maxCapacity:
        type: integer
      size:
        $ref: '#/definitions/model.Size'
      type:
        type: string
    type: object
  model.FeedingSchedule:
    properties:
      animalID:
        type: string
      feedingTime:
        type: string
      foodType:
        type: string
    type: object
  model.Food:
    properties:
      foodType:
        type: string
      name:
        type: string
    type: object
  model.Gender:
    enum:
    - male
    - female
    type: string
    x-enum-varnames:
    - Male
    - Female
  model.Size:
    properties:
      height:
        type: integer
      lenght:
        type: integer
      width:
        type: integer
    type: object
  model.Species:
    properties:
      animalType:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: This is a sample server for my API using Chi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: My Chi API
  version: "1.0"
paths:
  /animals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Animal'
            type: array
      summary: Получить всех животных по виду
      tags:
      - ZooStat
  /api/animals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Animal'
            type: array
      summary: Получить всех животных
      tags:
      - animals
    post:
      consumes:
      - application/json
      parameters:
      - description: Animal Data
        in: body
        name: animal
        required: true
        schema:
          $ref: '#/definitions/model.Animal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Animal'
      summary: Добавить животное
      tags:
      - animals
  /api/animals/{id}:
    delete:
      parameters:
      - description: Animal ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Удалить животное
      tags:
      - animals
    get:
      parameters:
      - description: Animal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Animal'
      summary: Получить животное по ID
      tags:
      - animals
  /api/schedules:
    delete:
      consumes:
      - application/json
      description: Delete schedule by animalID and time from repository
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete feeding schedule
      tags:
      - feeding_schedule
    get:
      consumes:
      - application/json
      description: Get feeding schedule by ID
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.FeedingSchedule'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get feeding schedule by ID
      tags:
      - feeding_schedule
    post:
      consumes:
      - application/json
      description: Adds a new feeding schedule for an animal with specified feeding
        time and food type
      parameters:
      - description: Feeding schedule information
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/controllers.AddScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Add a new feeding schedule
      tags:
      - feeding_schedule
  /zoostat/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Enclosure'
      summary: Получить все клетки
      tags:
      - ZooStat
swagger: "2.0"